ledger2beancount: quote "code" metadata entries
We have to quote the meta data value for the "code" key because
beancount interprets something like:

  code: AT180020040000010374623

as a commodity.  It will be seen as a quoted when quoted.  Since
ledger allows text as code (rather than just numbers) quoting
it seems light the right thing to do.  (Arguably, we could also
check first if it's an integer but this isn't done.)