process ledger file in stanzas
Previously the script assumed that transacations are separated by an
empty line, but this assumption is incorrect.  You can also have
transactions without any separation or separated by comma.  The same
applies to other ledger directives, such as account declarations.

It's better to process the ledger file stanza by stanza: we know that
everything beloning to a directive has to start with at least one space,
so read a whole stanza at once and process it.  This approach also
allows us to better know what we're processing.