add option to specify configuration file
Allow users to specify a config file as an argument in case they
need different config files for different options.

Note that Config::Onion is weird about the name of the file.  The
file has to end in .yml or .yaml but Config::Onion expects the
filename without any extension.

Fixes #63