handle block comments without indentation correctly
The example for block comments in the ledger manual suggests that such
comments have to be indented but this is not the case.  Something like
this is allowed in ledger:

comment
foo bar
end comment

The only requirement for block comments appears to be that they are
terminated by "end comment" or "end test" (regardless of whether they
were started with "comment" or "test").  Just "end", which I assumed
was allowed, is now valid, and neither are "!end comment" or "@end
comment".

Fixes #98