add conversion note when something is renamed
Add a conversion note when an account, commodity or metadata key is
renamed by ledger2beancount.  While this can be fairly noisy, it's
good to show all automatic renames so users can check if they are
happy with the result.  If they want to decrease the noise, they can
add an appropriate mapping to one of the `_map` configs.