partly revert: use mm/dd for short dates
I changed mm-dd to mm/dd for the short format because ledger doesn't
recognise mm/dd when you use --input-date-format (ledger issue #566).
However, Michael Budde pointed out that you can specify
--input-date-format multiple times and you can specify
--input-date-format %m-%d in addition to the full forma to make it
work.

Change back to mm-dd in most cases (and in the default config) since
this is more in line with the default of yyyy-mm-dd.