Add support for posting-level dates
In addition to effective dates, ledger has support for dates on the
posting-level.  Store this information as metadata.

The current beancount proposal allows for such posting-level dates.
It will be interesting though what to do if a posting has both a
primary and aux date since that can't be expressed with the beancount
proposal as far as I can tell.

Fixes #125