Handle metadata on all postings
There were some postings (for example those with cost or price
information) where metadata in a comment on the same line as
the posting was ignored.

If the metadata is on the same line as the posting ($defer = 1),
store them in @cur_txn_meta so we can print them after we print
the posting.