Handle simple implicit conversions
Implicit conversions are generally a bad idea but let's try to support
them.  This shows the limitations of the current approach of
ledger2beancount.  We really should be storing parsed lines in some
data structure to make more analysis easier.  Instead, I have to parse
the lines (which are now beancount!) again to get the info I need.

Fixes #116