Handle account names that contain brackets
ledger allows brackets in account names.  You just have to be careful
to distinguish between virtual postings and account names with
brackets in them.

Instead of updating the regex, I found it easier to parse the brackets
from virtual postings as part of the account name and strip them later.

Fixes #161