Handle transactions with a single posting
There are two scenarios when a transaction can have a single posting
(without the use of a bucket); in both cases, there must be nothing
to balance.

1) A transaction with a posting with an amount that is 0

2) A transaction with a posting with a balance assignment that is
the same as the current value of the account.

We have to avoid putting such transactions into beancount because
they are not valid there.

Fixes #182