Show transactions with only one posting
Since version 2.0, ledger2beancount deals with balance assertions in
transactions with a single posting.

But there's another use case of transactions with a single posting:
voided checkes.

For example,

    2020-04-06 (1052) VOID
        * Assets:Checking   $-0.00

Currently this transaction is silently ignored by ledger2beancount.

The better solution is to re-use the account name from the first posting
to add a second posting.  This is of course only possible if there's
an amount (which has to be zero, although we don't check for that since
ledger already rejects such invalid transactions).  If there's no amount,
we cannot add the transaction because there would be two postings without
amount, which is invalid.

Fixes #186