Don't convert meta-data on ignored virtual postings
If a virtual posting (which is ignored) is followed by lines consisting
of meta-data, we might end up with duplicate meta-data because that
meta-data will now be associated with the previous posting.

We have to skip meta-data for ignored virtual postings.  But the
problem is that there might be interesting comments that should be
preserved.  Unfortunately, it's not easy to distinguish between the
two without calling handle_comment(), which we don't want to do.
So just preserve them as comments.

Note that meta-data or comments on the line of the virtual posting
itself are always ignored.  I don't think it's worth handling this
corner case.

Fixes #189