Improve support for inline math
One challenge with inline math is that ledger allows features that
are not supported in beancount, specifically ledger allows you to
call functions (such as abs) in inline math.

The best way to figure out whether such complex inline math is
used without actually parsing the math is to get rid of all the
math operators and see what's left over.  If there's a single
term, that's the commodity we need.  If there are multiple terms,
beancount wouldn't be able to handle the inline math construct.

Fixes #192