Improve preservation of indentation
Slightly improve the preservation of indentation of postings,
in particular with amounts without leading zero (.10) and
inline math.

I originally wanted to go further than this, but the problem is
that different users may have different expectations, and really
this can get messy quickly and the real solution is probably to
just use bean-format (as documented).