Inherit leading whitespace when splitting comments
For postings like this:

    Assets:Test                        10.00 EUR ; :bar: test
    Equity:Opening-Balance            -10.00 EUR ;:bar: test

what should the leading whitespace be after we split out the
:bar: tag from the comment?

Previously we took what was right in front of "test", but
that doesn't make sense.  The comment "style" is defined by
what comes immediately after the comment marker ;.  So let's
use that whitespace for the comment.