Round calculated total if needed for `price==cost` comparison
We compare the total price and total cost to see if we can drop the
price information.  However, a total calculated from a per-unit lot
might have more precision than a given total.  Therefore, if a
total was provided, we have to round the calculated total for the
comparison to work.

Fixes #234