Add option to move posting-level tags to the transaction
Beancount doesn't have tags on the level of postings at the
moment.  We store these as metadata instead (with the key
"tags").  However, this isn't optimal since these are not
proper tags.  Therefore, add an option that moves posting-level
tags to the transaction itself.

There are pros and cons for both methods (documented in the
manual).

Fixes #241