Print warning if account in assertion has sub-accounts
Balance assertions in ledger and beancount have one major
difference: ledger takes the total of the specified account,
but beancount takes the total of that account *plus* all its
sub-accounts.  Beancount's behaviour is the same as hledger's
=* syntax.

If we see an assertion for an account that has a sub-account
somewhere, print a warning (unless the assertion was =*).

Fixes #243