Fix parsing of amounts with decimal comma in hledger mode
If an amount using a decimal comma is given in hledger mode
that has no commodity (i.e. it used the default commodity),
the comma would simply be stripped, leading to a wrong number
(e.g. 1,00 -> 100 instead of 1.00).

This was because we didn't store the original format of the
commodity definition, and the processing of amounts relies on
that format.

Fixes #259