Avoid incorrect collision warning in implicit conversions
In implicit conversions, we're adding the other commodity as a
price to make the conversion explicit.

The other commodity is already converted to a beancount commodity,
so if a ledger commodity requires a rename, we see both the ledger
commodity and now the beancount commodity, leading to a collision
warning.

However, that warning is bogus because the beancount commodity was
added by ledger2beancount and is not in the original ledger file.

Avoid that by setting the commodity directly instead of going through
get_amount() which involves map_commodity() which leads to the
collision warning by setting %ledger_commodities.

Fixes #265